Add files via upload
here is my attempt at the home learning. I forgot the week 1 home learning, so I did what I remembered. for Week 2 I made an attempt at the SRS there were some parts of the documentation I did not complete, for example: Product Perspective. I was not sure if this was just a repetition of all that I had wrote down previously or if it was something different.